moved Home file to root
